+++ eval/variables.go
@@ -1,11 +1,13 @@
 package eval
 
 const (
+	Backends         = "backends"
 	BackendRequest   = "backend_request"
 	BackendRequests  = "backend_requests"
 	BackendResponse  = "backend_response"
 	BackendResponses = "backend_responses"
 	BackendDefault   = "default"
+	State            = "state"
 	Body             = "body"
 	ClientRequest    = "request"
 	CTX              = "context"
@@ -14,6 +16,7 @@ const (
 	Environment      = "env"
 	FormBody         = "form_body"
 	Headers          = "headers"
+	Health           = "health"
 	HttpStatus       = "status"
 	ID               = "id"
 	JsonBody         = "json_body"